refactor: adjust the directory structure
