+++ src/languages/dtlog/dtlog.ts
