--- tsconfig.build.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "lib",
-    "module": "esnext",
-    "target": "ES6",
-    "declaration": true,
-    "jsx": "react",
-    "moduleResolution":"Node",
-    "keyofStringsOnly": true,
-    "allowSyntheticDefaultImports": true,
-    "experimentalDecorators": true,
-    "skipLibCheck": true
-  },
-  "include": ["src/components/editor"],
-  "exclude": ["node_modules", "build", "scripts", "**/__tests__/*.test.tsx",]
-}
