+++ tsconfig.json
@@ -1,41 +1,15 @@
 {
-    "compilerOptions": {
-        "noEmit": true, // Don't emit; allow Babel to transform files.
-        "outDir": "./dist/",
-        "baseUrl": ".",
-        "experimentalDecorators": true,
-        "noImplicitAny": false,
-        "esModuleInterop": true,
-        "allowJs": true,
-        "module": "commonjs",
-        "target": "es5",
-        "lib": [
-            "es6",
-            "es2016",
-            "dom"
-        ],
-        "jsx": "react",
-        "allowSyntheticDefaultImports": true,
-        "noUnusedLocals": false,
-        "importHelpers": true,
-        "paths": {},
-        "skipLibCheck": true,
-        "typeRoots": [
-            "node",
-            "node_modules/@types",
-            "./typings"
-        ],
-    },
-    "exclude": [
-        "node_modules",
-        "dist",
-        "lib",
-        "storybook-static",
-        "src/public/**/*"
-    ],
-    "include": [
-        "./**/*.ts",
-        "./**/*.tsx",
-        "src/**/*"
-    ]
+  "compilerOptions": {
+    "outDir": "lib",
+    "module": "esnext",
+    "target": "ES6",
+    "declaration": true,
+    "jsx": "react",
+    "moduleResolution":"Node",
+    "allowSyntheticDefaultImports": true,
+    "experimentalDecorators": true,
+    "skipLibCheck": true
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "build", "scripts", "lib"]
 }
